Add info about my experience with this tutorial
Throughout the completion of this project, I see how this can be a useful tool for collaboration between employees and project coordinators both remotely from anywhere in the world, or within the same building. This cuts down on the need for large meeting rooms, confusing video conferences, and other less organized means of collaboration and idea sharing.